+++ src/App.jsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import { BrowserRouter, Routes, Route, NavLink } from "react-router-dom";
-import TodoList from "./Components/TodoList";
-import AddTodoForm from "./Components/AddTodoForm";
+import TodoContainer from "./Components/TodoContainer";
 import styles from "./App.module.css"
 
 function Navigation() {
@@ -32,126 +31,6 @@ function Navigation() {
 }
 
 function App() {
-  const [todoList, setTodoList] = React.useState([]);
-  const [isLoading, setIsLoading] = React.useState(true);
-
-  async function fetchData() {
-    const url = `https://api.airtable.com/v0/${
-      import.meta.env.VITE_AIRTABLE_BASE_ID
-    }/${import.meta.env.VITE_TABLE_NAME}`;
-
-    const options = {
-      method: "GET",
-      headers: {
-        Authorization: `Bearer ${import.meta.env.VITE_AIRTABLE_API_TOKEN}`,
-      },
-    };
-
-    try {
-      const response = await fetch(url, options);
-
-      if (!response.ok) {
-        const message = `Error: ${response.status}`;
-        throw new Error(message);
-      }
-
-      const data = await response.json();
-
-      const todos = data.records.map((todo) => {
-        return {
-          id: todo.id,
-          title: todo.fields.title,
-        };
-      });
-
-      setTodoList(todos);
-      setIsLoading(false);
-    } catch (error) {
-      console.log("Fetch error:", error.message);
-      return null
-    }
-  }
-
-  React.useEffect(() => {
-    fetchData();
-  }, []);
-
-  async function addTodo(newTodoTitle) {
-
-    const url = `https://api.airtable.com/v0/${
-      import.meta.env.VITE_AIRTABLE_BASE_ID
-    }/${import.meta.env.VITE_TABLE_NAME}`;
-
-    const options = {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${import.meta.env.VITE_AIRTABLE_API_TOKEN}`,
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        records: [
-          {
-            fields: {
-              title: newTodoTitle,
-            },
-          },
-        ],
-      }),
-    };
-    
-    try {
-      const response = await fetch(url, options);
-      
-      if (!response.ok) {
-        throw new Error(`Error: ${response.status}`);
-      }
-
-      const data = await response.json();
-  
-      const newTodo = {
-        id: data.records[0].id,
-        title: data.records[0].fields.title,
-      };
-
-      setTodoList((prevTodoList) => [...prevTodoList, newTodo]);
-    } catch (error) {
-      console.log(error.message);
-      return null                          
-    }
-  }
-
-  async function removeTodo(id) {
-    const url = `https://api.airtable.com/v0/${
-      import.meta.env.VITE_AIRTABLE_BASE_ID
-    }/${import.meta.env.VITE_TABLE_NAME}/${id}`;
-
-    const options = {
-      method: "DELETE",
-      headers: {
-        Authorization: `Bearer ${import.meta.env.VITE_AIRTABLE_API_TOKEN}`,
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        records: [id],
-      }),
-    };
-
-    try {
-      const response = await fetch(url, options);
-
-      if (!response.ok) {
-        throw new Error(`Error: ${response.status}`);
-      }
-
-      const data = await response.json();
-      setTodoList((prevTodoList) =>
-        prevTodoList.filter((todo) => todo.id !== data.id)
-      );
-    } catch (error) {
-      console.log(error.message);
-      return null;
-    }
-  }
 
     return (
       <main className={styles.main}>
@@ -161,22 +40,7 @@ function App() {
             <Routes>
               <Route
                 path="/"
-                element={
-                  <div className={styles.headerAndList}>
-                    <h1 className={styles.header}>TODO LIST</h1>
-                    <div className={styles.formAndList}>
-                      <AddTodoForm onAddTodo={addTodo} />
-                      {isLoading ? (
-                        <p>Loading...</p>
-                      ) : (
-                        <TodoList
-                          todoList={todoList}
-                          onRemoveTodo={removeTodo}
-                        />
-                      )}
-                    </div>
-                  </div>
-                }
+                element= {<TodoContainer/>}
               />
               <Route
                 path="/new"
