+++ src/App.module.css
@@ -13,26 +13,6 @@
   min-width: 50%;
 }
 
-.headerAndList {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.header {
-  font-size: 65px;
-  color: #6f7083;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-}
-
-.formAndList {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
 .navContainer {
   display: flex;
   justify-content: center;
