+++ src/Components/TodoContainer.module.css
@@ -1,0 +1,19 @@
+.todoContainer {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+.header {
+  font-size: 65px;
+  color: #6f7083;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+}
+
+.formAndList {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
